# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IAdminSaveResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Name",
					"Email",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAdminSaveRequest": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"Name",
					"Email",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IadminLoginResponse": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAdminLoginRequest": {
				"properties": {
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"Email",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGroupMembers": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGroupSaveResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Members": {
						"items": {
							"$ref": "#/components/schemas/IGroupMembers"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"Name",
					"Members"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGroupSaveRequest": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Members": {
						"items": {
							"$ref": "#/components/schemas/IGroupMembers"
						},
						"type": "array"
					}
				},
				"required": [
					"Name",
					"Members"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGroupAddMemberResponse": {
				"properties": {
					"Members": {
						"items": {
							"$ref": "#/components/schemas/IGroupMembers"
						},
						"type": "array"
					}
				},
				"required": [
					"Members"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGroupAddMemberRequest": {
				"properties": {
					"GroupId": {
						"type": "string"
					},
					"UserId": {
						"type": "string"
					}
				},
				"required": [
					"GroupId",
					"UserId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGroupGetResponse": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"IGroupGetRequest": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGroupDeleteRequest": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGroupSearchWordResponse": {
				"properties": {
					"User": {
						"type": "string"
					},
					"Message": {
						"type": "string"
					},
					"Time": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"User",
					"Message",
					"Time"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGroupSearchWordRequest": {
				"properties": {
					"Word": {
						"type": "string"
					},
					"GroupId": {
						"type": "string"
					}
				},
				"required": [
					"Word",
					"GroupId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserGroups": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserSaveUpdateResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"Groups": {
						"items": {
							"$ref": "#/components/schemas/IUserGroups"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"Name",
					"Email",
					"Password",
					"Groups"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserSaveRequest": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					}
				},
				"required": [
					"Name",
					"Email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserUpdateRequest": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"Groups": {
						"items": {
							"$ref": "#/components/schemas/IUserGroups"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"Name",
					"Email",
					"Password",
					"Groups"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserDeleteRequest": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGroupConversationResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Members": {
						"items": {
							"$ref": "#/components/schemas/IGroupMembers"
						},
						"type": "array"
					},
					"Conversation": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Name",
					"Members",
					"Conversation"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IConversationSendMessageRequest": {
				"properties": {
					"UserId": {
						"type": "string"
					},
					"GroupId": {
						"type": "string"
					},
					"Message": {
						"type": "string"
					}
				},
				"required": [
					"UserId",
					"GroupId",
					"Message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Final-Task",
		"version": "1.0.0",
		"description": "this is the description of the file",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/saveAdmin": {
			"post": {
				"operationId": "SaveAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAdminSaveResponse"
								}
							}
						}
					}
				},
				"description": "Create New Admin",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAdminSaveRequest"
							}
						}
					}
				}
			}
		},
		"/admin/loginAdmin": {
			"post": {
				"operationId": "AdminLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IadminLoginResponse"
								}
							}
						}
					}
				},
				"description": "Admin Login",
				"tags": [
					"Admin Login"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAdminLoginRequest"
							}
						}
					}
				}
			}
		},
		"/group/saveGroup": {
			"post": {
				"operationId": "SaveGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGroupSaveResponse"
								}
							}
						}
					}
				},
				"description": "Save New Group",
				"tags": [
					"Groups"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGroupSaveRequest"
							}
						}
					}
				}
			}
		},
		"/group/addMemberToGroup": {
			"put": {
				"operationId": "AddMemberToGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGroupAddMemberResponse"
								}
							}
						}
					}
				},
				"description": "Add Member to Group",
				"tags": [
					"Groups"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGroupAddMemberRequest"
							}
						}
					}
				}
			}
		},
		"/group/getGroup": {
			"post": {
				"operationId": "GetGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGroupGetResponse"
								}
							}
						}
					}
				},
				"description": "Get Details of  Group",
				"tags": [
					"Groups"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGroupGetRequest"
							}
						}
					}
				}
			}
		},
		"/group/deleteGroup": {
			"delete": {
				"operationId": "DeleteGroup",
				"responses": {
					"200": {
						"description": "Group Deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGroupGetResponse"
								}
							}
						}
					}
				},
				"description": "Delete Group",
				"tags": [
					"Groups"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGroupDeleteRequest"
							}
						}
					}
				}
			}
		},
		"/group/searchWord": {
			"post": {
				"operationId": "SearchWord",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IGroupSearchWordResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Search Words in Group",
				"tags": [
					"Search Word"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGroupSearchWordRequest"
							}
						}
					}
				}
			}
		},
		"/user/saveUser": {
			"post": {
				"operationId": "SaveUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserSaveUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Regisater User",
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserSaveRequest"
							}
						}
					}
				}
			}
		},
		"/user/updateUser": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserSaveUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Update Existing User",
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/user/deleteUser": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "User Deleted Successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Delete Existing User",
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserDeleteRequest"
							}
						}
					}
				}
			}
		},
		"/user/sendMessage": {
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGroupConversationResponse"
								}
							}
						}
					}
				},
				"description": "Sending Message to a group",
				"tags": [
					"Conversation"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IConversationSendMessageRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"tags": [
		{
			"name": "Admin Operations",
			"description": "Here all the CRUD operations of the Admin are given",
			"externalDocs": {
				"description": "Find more about Admin",
				"url": "http://admin.io"
			}
		},
		{
			"name": "Admin Login",
			"description": "This is Admin Login only"
		},
		{
			"name": "Admin",
			"description": "Here We can save new Admin"
		},
		{
			"name": "User",
			"description": "Here admin can do operations about users"
		},
		{
			"name": "Search Word",
			"description": "Here admin can search word in group messages"
		},
		{
			"name": "Groups",
			"description": "Here admin can do operations about groups"
		},
		{
			"name": "Conversation",
			"description": "Here suer can send messages to the user"
		}
	]
}