# src/types/requests/group.request.ts
import { IGroupMembers } from '../documents/group.document';

export interface IGroupSaveRequest {
	Name: string;
	Members: IGroupMembers[];
}
export interface IGroupAddMemberRequest {
	GroupId: string;
	UserId: string;
}
export interface IGroupDeleteRequest {
	_id: string;
}

export interface IGroupGetRequest {
	_id: string;
}


export interface IUpdateConversation {
	GroupId: string;
	ConversationId: string;
}
export interface IGroupSearchWordRequest {
	Word: string;
	GroupId: string;
}